Add renaming of modules
