+++ backend/pkg/auth/auth_config.go
@@ -1,7 +1,7 @@
 package auth
 
 import (
-	dbstore "github.com/grafana/simple-datasource-backend/pkg/db"
+	"github.com/grafana/simple-datasource-backend/pkg/dbstore"
 )
 
 type AuthConfig struct {
