+++ backend/pkg/auth/email_config.go
@@ -3,7 +3,7 @@ package auth
 import (
 	"database/sql"
 
-	dbstore "github.com/grafana/simple-datasource-backend/pkg/db"
+	"github.com/grafana/simple-datasource-backend/pkg/dbstore"
 )
 
 type EmailConfig struct {
