--- backend/pkg/server/report_recipient.go
@@ -1,1 +1,0 @@
-package server
